BREAKING: move flat configs to ./flat
