+++ package.json
@@ -31,7 +31,6 @@
   },
   "files": [
     "*.js",
-    "configs",
     "legacy",
     "rules",
     "LICENSE"
